🎨 Create seperate JS files for header, amin and footer components
