+++ lib/footerElement.js
@@ -1,0 +1,9 @@
+import { createElement } from "./elements";
+
+export function createFooterElement() {
+  const footerElement = createElement("footer", { className: "footer" }, [
+    "footer",
+  ]);
+
+  return footerElement;
+}
