+++ lib/mainElement.js
@@ -1,0 +1,9 @@
+import { createElement } from "./elements";
+
+export function createMainElement() {
+  const mainElement = createElement("main", { className: "main" }, [
+    "in development...",
+  ]);
+
+  return mainElement;
+}
